Fix "rush add" tests.
