+++ apps/rush-lib/src/cli/actions/test/addRepo/a/package.json
